fix sign-compare and format warning on i386
i386 builds where broken due to an incorrect sign-compare
and the use of a 64bit specific format string

+++ common/rfc6287.h
@@ -31,7 +31,7 @@ typedef struct ocra_suite_struct {
 	/* DataInput */
 	int	 flags;
 	enum fmt Q_fmt;
-	size_t	 Q_l;
+	int	 Q_l;
 	enum alg P_alg;
 	size_t	 S_l;
 	int	 T_step;
